Integrate `_idl_stuff` into `_memberspec`.
`_idl_stuff.py` each line blame:
55c56cc1
3579c214
e1d61858
dbe4fa1b
