+++ comtypes/__init__.py
@@ -39,8 +39,6 @@ from comtypes import patcher  # noqa
 from comtypes._npsupport import interop as npsupport  # noqa
 from comtypes._tlib_version_checker import _check_version  # noqa
 from comtypes._py_instance_method import instancemethod  # noqa
-from comtypes._idl_stuff import defaultvalue, helpstring, dispid  # noqa
-from comtypes._idl_stuff import STDMETHOD, DISPMETHOD, DISPPROPERTY, COMMETHOD  # noqa
 
 _all_slice = slice(None, None, None)
 
@@ -168,6 +166,19 @@ com_interface_registry = {}
 # allows to find coclasses by guid strings (clsid)
 com_coclass_registry = {}
 
+
+################################################################
+# IDL stuff
+
+from comtypes._memberspec import (  # noqa
+    COMMETHOD,
+    DISPMETHOD,
+    DISPPROPERTY,
+    STDMETHOD,
+    defaultvalue,
+    dispid,
+    helpstring,
+)
 
 ################################################################
 # IUnknown, the root of all evil...
