+++ comtypes/_post_coinit/unknwn.py
@@ -8,9 +8,8 @@ from typing import Optional
 from typing import List, Type
 
 from comtypes import _ole32_nohresult, com_interface_registry
-from comtypes._idl_stuff import STDMETHOD
 from comtypes._memberspec import ComMemberGenerator, DispMemberGenerator
-from comtypes._memberspec import _ComMemberSpec, _DispMemberSpec
+from comtypes._memberspec import STDMETHOD, _ComMemberSpec, _DispMemberSpec
 from comtypes._post_coinit import _cominterface_meta_patcher as _meta_patch
 from comtypes._post_coinit.guid import GUID
 from comtypes._py_instance_method import instancemethod
